refactor(jpath): dirs.go
